Use `heroku local` instead of `foreman start`
This commit changes the instructions that are printed once a raygun
project is generated, to bring it in line with latest practices.

First, `bin/rake` is preferable to `rake` (i.e. if `./bin` is not in the
`$PATH`). The binstub ensures that the right version of rake is used,
and also has performance benefits due to spring. This also fixes
problems like this:

```
$ rake
rake aborted!
Gem::LoadError: You have already activated rake 12.3.0, but your Gemfile requires rake 12.2.1. Prepending `bundle exec` to your command may solve this.
```

Second, `heroku local` is preferable to `foreman start`, as explained
here: https://github.com/carbonfive/raygun-rails/pull/99

Finally, as a matter of style and brevity, use `bin/<command>` instead
of `./bin/<command>`.