Remove URI.encode, which is deprecated in Ruby 2.7
This fixes the following warning when running raygun under Ruby 2.7.0:

```
warning: URI.escape is obsolete
```

We don't actually need to escape anything, because in practice a branch
name should be simply appended to the GitHub URL as-is. For example, if
the branch is `dependabot/bundler/pg-1.2.2`, then this URL is completely
valid:

```
https://github.com/carbonfive/raygun-rails/archive/dependabot/bundler/pg-1.2.2.tar.gz
```

I confirmed this by testing locally.