Run c5-conventions/rubocop on this project
Problem
-------

Shouldn't we eat our own dog food?

This repo was way out of sync with our current rubocop conventions.

Solution
--------

Bring it up to speed.

Because we don't have `rubocop` in the project, this can be seen as
a one-time cleanup.  We can decide later if we want it more persistent.
But for now I did the following:

* Copy `.rubocop.yml` from `carbonfive/c5-conventions`
  ```bash
  curl https://raw.githubusercontent.com/carbonfive/c5-conventions/master/rubocop/rubocop.yml > .rubocop.yml
  ```
* Locally add `rubocop` and `rubocop-performance` gems
  ```bash
  gem install rubocop
  gem install rubocop-performance

  ```
* run it
  ```bash
  rubocop -a
  ```