Problem --------
Our connection to the rubocop.yml file in https://github.com/carbonfive/c5-conventions seems fragile.

Updates to that repo will affect all consumers out in the world (who we don't know) and may cause issues with their app. We've already seen it affect 1 client project in house.

Solution
---------

By default, pull the `rubocop.yml` from `c5-conventions` at the time the
app is constructed and pin it.  This will disconnect the constructed app
from the possibly dynamic rubocop file.

Additionally, we add some commentary to the top of the file to indcate
the SHA which was the original pull and a note to suggest upgrading at
the developers' discretion.

Changes
-------

* by default, fetch the rubocop file from github and write it to the
newly created target app
* add rspec
* add a test for this new functionality (the pull of the rubocop from GitHub)

Fixes #146