Add bin/setup script and development instructions
As is tradition for Rails and gem codebases, add a `bin/setup` script
to bootstrap a developer's local environment. For raygun this is just a
matter of running `bundle install`. I took this script from the standard
gem template published by the bundler project:

https://github.com/rubygems/bundler/blob/203d02c3670083a95733afc1a9d7a870d1ae1c34/lib/bundler/templates/newgem/bin/setup.tt

Also modify the README to mention `bin/setup` and explain how to run
tests and rubocop.