Rename master branch to "main" (#157)
This commit updates the raygun docs to reflect "main" as the new default
branch name for this repo (replacing "master").

Additionally, when raygun "zaps" a new project, it will now initialize
the git repo of the new project with "main" as the default branch name.
This is done by calling:

```
git checkout -b main
```

After the repo is initialized. This works because an freshly init-ed
repo does not have any branches, so the act of creating a branch
implicitly makes it the default.