Document how to add React and Typescript
Apps generated with raygun don't include React or Typescript out of the
box, but these are often a natural next step.

Adding React and Typescript to Rails is simple with webpacker, but this
is not widely known. Help developers who may be adding React or
Typescript to a webpacker project for the first time by documenting it
in the raygun README.

Also call out React in the next steps that are printed by raygun.