Remove explicit .ruby-version setting (#168)
raygun should work with any version of ruby >= 2.5 (as specified in the
gemspec), and is tested in CI with 2.5, 2.6, 2.7, and 3.0. So it doesn't
really make sense to hardcode the development environment to a specific
ruby version.

More importantly, when the raygun repo itself is pinned to a specific
version of ruby, running raygun from a local copy of the repo leads to
some confusing behavior. There are two pinned versions of ruby at play:
`raygun/.ruby-version` and `raygun-rails/.ruby-version`. It is not clear
which should take precedence.

See additional discussion in #167.

This commit removes this confusion by removing the `.ruby-version` file
from this repo. The gemspec will still enforce a compatible ruby
version, so the version file was somewhat redundant.

I also removed `.ruby-gemset` because I don't think that was being used.

Fixes #167.